% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telechargement_donnes.R
\name{ACC_ressources}
\alias{ACC_ressources}
\title{Ensemble des ressources du jeu de données Accident}
\usage{
ACC_ressources()
}
\value{
Une table de toutes les ressources incluant l'url pour les télécharger.
}
\description{
Le package \code{BARIS} permet de récupérer les ressources disponibles sur le site
\href{https://data.gouv.fr/}{data.gouv.fr} avec la fonction \link[BARIS:BARIS_resources]{BARIS::BARIS_resources}. Seules sont renvoyées
les ressources au format \code{csv}.
}
\examples{
\dontrun{
ressources <- ACC_ressources()
}
}
\seealso{
\link[BARIS:BARIS_resources]{BARIS::BARIS_resources}
}
