% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{caracteristiques}
\alias{caracteristiques}
\title{Caracteristiques}
\format{
Un data.frame avec les variables :
...
}
\usage{
caracteristiques
}
\description{
Ajouter la description de la table caracteristiques
}
\keyword{datasets}
