% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_accidents.R
\name{create_table_accidents}
\alias{create_table_accidents}
\title{Table unique des accidents}
\usage{
create_table_accidents(caracteristiques, usagers, lieux, vehicules)
}
\arguments{
\item{caracteristiques}{data.frame obtenu par \code{\link[=dl_caracteristiques]{dl_caracteristiques()}} \verb{\%>\%} \code{\link[=recod_caracteristiques]{recod_caracteristiques()}}}

\item{usagers}{data.frame obtenu par \code{\link[=dl_usagers]{dl_usagers()}} \verb{\%>\%} \code{\link[=recod_usagers]{recod_usagers()}}}

\item{lieux}{data.frame obtenu par \code{\link[=dl_lieux]{dl_lieux()}} \verb{\%>\%} \code{\link[=recod_lieux]{recod_lieux()}}}

\item{vehicules}{data.frame obtenu par \code{\link[=dl_vehicules]{dl_vehicules()}} \verb{\%>\%} \code{\link[=recod_vehicules]{recod_vehicules()}}}
}
\value{
Un tibble des différentes tables jointes
}
\description{
Table unique des accidents
}
