% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telechargement_donnes.R
\name{dl_data}
\alias{dl_data}
\alias{dl_caracteristiques}
\alias{dl_lieux}
\alias{dl_usagers}
\alias{dl_vehicules}
\title{Telechargement des donnees}
\usage{
dl_caracteristiques(.annee = 2008:2019)

dl_lieux(.annee = 2008:2019)

dl_usagers(.annee = 2008:2019)

dl_vehicules(.annee = 2008:2019)
}
\arguments{
\item{.annee}{Un entier ou un vecteur d'entiers pour l'annee souhaitee}
}
\value{
Un \code{tibble}
}
\description{
Ces fonctions permettent de telecharger pour les 4 differentes ressources,
les donnees pour differentes annees.
Les ressources sont :
}
\details{
\itemize{
\item CARACTERISTIQUES
\item LIEUX
\item VEHICULES
\item USAGERS
}
}
\section{Functions}{
\itemize{
\item \code{dl_caracteristiques}: Telechargement des donnees

\item \code{dl_lieux}: Telechargement des donnees

\item \code{dl_usagers}: Telechargement des donnees

\item \code{dl_vehicules}: Telechargement des donnees
}}

\examples{
# Caracteristiques des accidents en 2019

\dontrun{
caracteristiques <- dl_caracteristiques()
usagers <- dl_usagers()
vehicules <- dl_vehicules()
lieux <- dl_lieux()
}

}
