% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_pour_modele.R
\name{mutate_pour_modele}
\alias{mutate_pour_modele}
\title{Modification des variables}
\usage{
mutate_pour_modele(accidents, population)
}
\arguments{
\item{accidents}{table accidents obtenue via \code{\link[=create_table_accidents]{create_table_accidents()}}}

\item{population}{table population obtenue via \code{data("population", package = "ACC")}}
}
\value{

}
\description{
Cette fonction permet d'appliquer des modifications \emph{finales} à la table
accidents comme l'ajout d'une variable supplémentaire sur la taille du
département, le regroupement de certaines modalités trop détaillées sur les
équipements de sécurité, le regroupement de l'âge en classes d'âges et
le regroupement de l'heure de l'accident en heure de pointe (matin et soir) et
les autres heures.
}
\details{
Ces regroupements ont pu être mis en évidence suite à l'exploration des données,
et nous ont parus intéressants.
}
