% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhanesSample.R
\docType{data}
\name{nhanesSample}
\alias{nhanesSample}
\title{NHANES data}
\format{A data frame with 184 rows and 4 variables, for 100 randomly sampled individuals in the NHANES 2003-04. This consists physical activity data for 100 individuals.}
\usage{
data(nhanesSample)
}
\description{
NHANES 2003-04 data
}
\details{
\itemize{
  \item ID: Individual's ID
  \item Time: Time in minutes for 7 observation days.
  \item Min: Minutes moderate to vigorous physical activity in 10 minute bouts.
  \item Female: Binary indicator (1=Female, 0=Male).
}
}
\keyword{datasets}

