% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCounts.R
\name{readCounts}
\alias{readCounts}
\title{Reads counts data in .dat, .agd, or .csv format for Actigraph GT1M and GT3X
devices}
\usage{
readCounts(
  filename,
  dateformat = "\%y-\%b-\%d",
  timeformat = "\%H:\%M:\%S",
  timezone = "GMT"
)
}
\arguments{
\item{filename}{Speficy full file path and file name. e.g. C:/mydata.dat or}

\item{dateformat}{format for dates}

\item{timeformat}{format for times}

\item{timezone}{time zone of the data captured
C:/mydata.csv}
}
\value{
For uni-axial accelerometer (GT1M), two columns are returned, consisting of:
[TimeStamp,Counts] For tri-axial accelerometer (GT3X), four columns are
returned, consisting of: [TimeStamp,x,y,z]
}
\description{
Reads counts data in .dat, .agd, or .csv format for Actigraph GT1M and GT3X
devices. Device type and epoch is automatically detected and reported in the
console.
}
\examples{


##
## A example to read counts data
##
\dontrun{
accData1 <- readCounts("C:/mydata.agd")
accData1 <- readCounts("C:/mydata.dat")
accData2 <- readCounts("C:/mydata.csv")
}



}
\author{
Jaejoon Song <jjsong2@mdanderson.org>
}
\keyword{accelerometer}
