% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage-selection.R
\name{recod_caracteristiques}
\alias{recod_caracteristiques}
\title{Recodage et sélection de la table caracteristiques}
\usage{
recod_caracteristiques(caracteristiques)
}
\arguments{
\item{caracteristiques}{Table des caractéristiques "brute"}
}
\value{
Un \code{data.frame} des caractéristiques avec une sélection de variables
et un recodage de certaines variables avec les libellés correspondants
}
\description{
Recodage et sélection de la table caracteristiques
}
\examples{
\dontrun{
caracteristiques <- dl_caracteristiques()
caracteristiques <- recod_caracteristiques(caracteristiques)
}
}
