% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage-selection.R
\name{recod_lieux}
\alias{recod_lieux}
\title{Recodage et sélection de la table lieux}
\usage{
recod_lieux(lieux)
}
\arguments{
\item{lieux}{Table des lieux d'accidents "brute"}
}
\value{
Un \code{data.frame} des lieux avec une sélection de variables
et un recodage de certaines variables avec les libellés correspondants
}
\description{
Recodage et sélection de la table lieux
}
\examples{
\dontrun{
 lieux <- dl_caracteristiques()
 lieux <- recod_caracteristiques(lieux)
}
}
