% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage-selection.R
\name{recod_usagers}
\alias{recod_usagers}
\title{Recodage et sélection de la table usagers}
\usage{
recod_usagers(usagers)
}
\arguments{
\item{usagers}{Table des usagers d'accidents "brute"}
}
\value{
Un \code{data.frame} des usagers avec une sélection de variables
et un recodage de certaines variables avec les libellés correspondants
}
\description{
Recodage et sélection de la table usagers
}
\examples{
\dontrun{
 usagers <- dl_usagers()
 usagers <- recod_usagers(usagers)
}
}
