% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodage-selection.R
\name{recod_vehicules}
\alias{recod_vehicules}
\title{Recodage et sélection de la table vehicules}
\usage{
recod_vehicules(vehicules)
}
\arguments{
\item{vehicules}{Table des vehicules d'accidents "brute"}
}
\value{
Un \code{data.frame} des vehicules avec une sélection de variables
et un recodage de certaines variables avec les libellés correspondants
}
\description{
Recodage et sélection de la table vehicules
}
\examples{
\dontrun{
 vehicules <- dl_vehicules()
 vehicules <- recod_vehicules(vehicules)
}
}
