% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aeexfit.R
\name{summary.aeexfit}
\alias{summary.aeexfit}
\title{Summary method for aeexfit objects}
\usage{
\method{summary}{aeexfit}(object, digits = 3, dig.tst = 2, ...)
}
\arguments{
\item{object}{A \code{aeexfit} object}

\item{digits}{Minimum number of significant digits to be used for most
numbers.}

\item{dig.tst}{Minimum number of significant digits for the test statistics}

\item{...}{...}
}
\value{
Estimated parameters for aeefit objects
}
\description{
Prints estimated parameters for \code{aeexfit} object
}
\examples{

##Will put an example here
}
\author{
Jaejoon Song <jjsong2@mdanderson.org>
}
\keyword{summary.aeexfit}
