% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicules}
\alias{vehicules}
\title{véhicules}
\format{
Un data.frame avec avec 504 834 lignes et 8 variables
\itemize{
\item \strong{Num_Acc} : Identifiant de l’accident
\item \strong{id_véhicule} : Identifiant unique du véhicule
\item \strong{num_veh} : Identifiant du véhicule repris pour chacun des usagers occupant ce véhicule
\item \strong{catv} : Categorie du véhicule
\itemize{
\item \code{Bicyclette}
\item \verb{2 roues motorisé}
\item \verb{VL seul}
\item \verb{VU seul}
\item \verb{Poids lourd}
\item \verb{Engin agricole}
\item \code{Bus}
\item \code{Train}
\item \code{Tramway}
\item \code{Quad}
\item \code{Autre}
}
\item \strong{obs} : Obstacle fixe heurté
\itemize{
\item \code{Non}
\item \code{Oui}
\item \verb{Non Renseigné}
}
\item \strong{obsm} : Obstacle mobile heurté
\itemize{
\item \code{Aucun}
\item \code{Piéton}
\item \code{Véhicule}
\item \code{Animal}
\item \code{Autre}
}
\item \strong{choc} : Point de choc initial
\itemize{
\item \code{Aucun}
\item \code{Avant}
\item \code{Arrière}
\item \code{Côté}
\item \verb{Chocs multiples (tonneaux)}
\item \verb{Non Renseigné}
}

...
}
}
\usage{
vehicules
}
\description{
véhicules
}
\keyword{datasets}
