% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\docType{data}
\name{weight}
\alias{weight}
\title{Sample weights for the subsample included in data(NHANES)}
\format{A data frame with 2763 rows and 2 variables}
\usage{
data(weight)
}
\description{
Sample weights for the subsample included in data(NHANES).
}
\details{
\itemize{
  \item ID: Individual's ID
  \item weight: Sample weights for the subsample included in data(NHANES)
}
}
\references{
National Cancer Institute. SAS Programs for analyzing NHANES 2003-2004 accelerometer data; [cited 2015 December 28]. Available from: http://appliedresearch.cancer.gov/nhanes_pam.
}
\keyword{datasets}

