% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vehicules}
\alias{vehicules}
\title{Vehicules}
\format{
Un data.frame avec les variables :
...
}
\usage{
vehicules
}
\description{
Ajouter la description de la table vehicules
}
\keyword{datasets}
